// src/enums/route.ts
export enum ERouter {
    DEFAULT = "/",
    LOGIN = "/auth/login",
    DASHBOARD = "/dashboard",
    DASHBOARD_OVERVIEW = "/dashboard/overview",
    DASHBOARD_USERS = "/dashboard/users",
    DASHBOARD_USER_ID = "/dashboard/users/:userId",
    DASHBOARD_CATEGORIES = "/dashboard/categories",
    DASHBOARD_CATEGORY_ID = "/dashboard/categories/:categoryId",
    DASHBOARD_PRODUCTS = "/dashboard/products",
    DASHBOARD_DOCUMENTS = "/dashboard/documents",
    DASHBOARD_CONTENT_PAGES = "/dashboard/content-pages",
    DASHBOARD_SETTINGS = "/dashboard/settings",
    NOTFOUND = "*",
}